refactor exposeAngularComponent to support both standalone or reuse existing injector
